+++ examples/07_serve_static.rs
@@ -33,20 +33,20 @@ fn main() {
         match fs::read(&path) {
             // If its found send it as response
             // We are setting the Content-Type header with the file extension through a match expression
-            Ok(content) => Response::new_raw(
-                200,
-                content,
-                vec![Header::new("Content-Type", get_type(&path))],
-            ),
+            Ok(content) => Response::new()
+                .status(200)
+                .bytes(content)
+                .header(Header::new("Content-Type", get_type(&path))),
 
             // If not read and send 404.html
             // If that file is not found, fallback to sending "Not Found :/"
-            Err(_) => Response::new_raw(
-                404,
-                fs::read(format!("{}/404.html", STATIC_DIR))
-                    .unwrap_or_else(|_| "Not Found :/".as_bytes().to_owned()),
-                vec![Header::new("Content-Type", "text/html")],
-            ),
+            Err(_) => Response::new()
+                .status(404)
+                .bytes(
+                    fs::read(format!("{}/404.html", STATIC_DIR))
+                        .unwrap_or_else(|_| "Not Found :/".as_bytes().to_owned()),
+                )
+                .header(Header::new("Content-Type", "text/html")),
         }
     });